requirements.txt
streamlit
requests
ultralytics
opencv-python-headless
numpy
python-dotenv

# pip install -r requirements.txt